// Website-Aplication/src/types/typeOrder.ts
export enum statusOrder {
    Processing = "Chờ xử lý",
    Repairing = "Đang sửa chữa",
    WaitForPay = "Chờ thanh toán",
    Finished = "Hoàn thành",
    Paid = "Đã thanh toán",

    Confirm = "Đã xác nhận",
    Transport = "Đang giao",
    Canceled = "Đã hủy",
}

// export enum typeOrder{
//     Purchase = "Đơn mua hàng",
//     Repair = "Đơn sửa chữa",
// }